Highlighted link to ALLSorts v1
ALLSorts V1 is linked to this repository in "The application of RNA sequencing for the diagnosis and genomic classification of pediatric acute lymphoblastic leukemia". This version of ALLSorts could be considered a "version 2" and hence now occupies this URL.

This was highlighted to me in issue #5. Hopefully this will reduce confusion.